Move integration tests into separate modules
