--- tests/health_check.rs
@@ -1,190 +1,0 @@
-use std::net::TcpListener;
-
-use once_cell::sync::Lazy;
-use sqlx::{Connection, Executor, PgConnection, PgPool};
-use uuid::Uuid;
-
-use email_newsletter::configuration::{get_configuration, DatabaseSettings};
-use email_newsletter::email_client::EmailClient;
-use email_newsletter::telemetry::{get_tracing_subscriber, init_subscriber};
-
-// ensure that the tracing stack is only initialized once
-static TRACING: Lazy<()> = Lazy::new(|| {
-    if std::env::var("TEST_LOG").is_ok() {
-        let subscriber = get_tracing_subscriber("test", "debug", std::io::stdout);
-        init_subscriber(subscriber);
-    } else {
-        let subscriber = get_tracing_subscriber("test", "debug", std::io::sink);
-        init_subscriber(subscriber);
-    }
-});
-
-// A struct holding data needed to access a test version of our application
-pub struct TestApp {
-    pub address: String,
-    pub connection_pool: PgPool,
-}
-
-// Spawns an app inside a future and returns the configured TestApp.
-async fn spawn_app() -> TestApp {
-    Lazy::force(&TRACING);
-
-    let listener = TcpListener::bind("127.0.0.1:0").expect("Failed to bind a random port");
-    let port = listener.local_addr().unwrap().port();
-    let address = format!("http://127.0.0.1:{}", port);
-
-    let mut configuration = get_configuration().expect("Failed to read configuration.");
-    configuration.database.database_name = Uuid::new_v4().to_string();
-
-    let connection_pool = configure_database(&configuration.database).await;
-
-    let sender_email = configuration
-        .email_client
-        .sender()
-        .expect("Invalid sender email address.");
-    let email_client = EmailClient::new(
-        configuration.email_client.base_url,
-        sender_email,
-        configuration.email_client.authorization_token,
-        std::time::Duration::from_millis(100),
-    );
-
-    let server = email_newsletter::startup::run(listener, connection_pool.clone(), email_client)
-        .expect("Failed to bind address");
-    let _ = tokio::spawn(server);
-    TestApp {
-        address,
-        connection_pool,
-    }
-}
-
-// Configures a test database, running all migrations, and then returning the connection pool handle
-// needed to use the test database.
-async fn configure_database(config: &DatabaseSettings) -> PgPool {
-    let mut connection = PgConnection::connect_with(&config.without_db())
-        .await
-        .expect("Failed to connect to postgres.");
-
-    connection
-        .execute(format!(r#"CREATE DATABASE "{}";"#, config.database_name).as_str())
-        .await
-        .expect("Failed to create database");
-
-    let connection_pool = PgPool::connect_with(config.with_db())
-        .await
-        .expect("Failed to connect to postgres.");
-
-    sqlx::migrate!("./migrations")
-        .run(&connection_pool)
-        .await
-        .expect("Failed to migrate the database");
-
-    connection_pool
-}
-
-#[tokio::test]
-async fn health_check_responds_200() {
-    // arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-
-    // act
-    let response = client
-        .get(&format!("{}/health_check", &test_app.address))
-        .send()
-        .await
-        .expect("Failed to execute request");
-
-    // assert
-    assert!(response.status().is_success());
-    assert_eq!(Some(0), response.content_length());
-}
-
-#[tokio::test]
-async fn subscribe_with_valid_form_data_returns_200() {
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-    let body = "name=le%20guin&email=ursula_le_guin%40gmail.com";
-
-    // act
-    let response = client
-        .post(&format!("{}/subscriptions", &test_app.address))
-        .header("Content-Type", "application/x-www-form-urlencoded")
-        .body(body)
-        .send()
-        .await
-        .expect("Failed to execute request");
-
-    // assert
-    assert_eq!(200, response.status().as_u16());
-
-    let saved_subscriber = sqlx::query!("SELECT email, name FROM subscriptions")
-        .fetch_one(&test_app.connection_pool)
-        .await
-        .expect("Failed to fetch saved subscription.");
-
-    assert_eq!(saved_subscriber.email, "ursula_le_guin@gmail.com");
-    assert_eq!(saved_subscriber.name, "le guin");
-}
-
-#[tokio::test]
-async fn subscribe_with_missing_form_data_returns_400() {
-    // arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-    let test_cases = vec![
-        ("name=le%20guin", "missing email"),
-        ("email=ursula_le_guin%40gmail.com", "missing name"),
-        ("", "missing both name and email"),
-    ];
-
-    for (invalid_body, error_message) in test_cases {
-        // act
-        let response = client
-            .post(&format!("{}/subscriptions", &test_app.address))
-            .header("Content-Type", "application/x-www-form-urlencoded")
-            .body(invalid_body)
-            .send()
-            .await
-            .expect("Failed to execute request");
-
-        // assert
-        assert_eq!(
-            400,
-            response.status().as_u16(),
-            "API did not fail with 400 when payload was {}",
-            error_message
-        )
-    }
-}
-
-#[tokio::test]
-async fn subscribe_with_invalid_fields_returns_400() {
-    // arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-    let test_cases = vec![
-        ("name=&email=test%40email.com", "empty name"),
-        ("name=test&email=", "empty email"),
-        ("name=test&email=invalid-email", "invalid email"),
-    ];
-
-    for (invalid_body, error_message) in test_cases {
-        // act
-        let response = client
-            .post(&format!("{}/subscriptions", &test_app.address))
-            .header("Content-Type", "application/x-www-form-urlencoded")
-            .body(invalid_body)
-            .send()
-            .await
-            .expect("Failed to execute request");
-
-        // assert
-        assert_eq!(
-            400,
-            response.status().as_u16(),
-            "API did not fail with 400 when payload was {}",
-            error_message
-        )
-    }
-}
